+++ packages/snap-preact/components/src/components/Organisms/RecommendationBundle/RecommendationBundle.stories.tsx
@@ -148,6 +148,17 @@ export default {
 			},
 			control: { type: 'text' },
 		},
+		hideSeed: {
+			description: 'Hide/show seed result',
+			defaultValue: false,
+			table: {
+				type: {
+					summary: 'boolean',
+				},
+				defaultValue: { summary: false },
+			},
+			control: { type: 'boolean' },
+		},
 		vertical: {
 			description: 'set the recommendation to render vertically',
 			defaultValue: false,
@@ -193,6 +204,19 @@ export default {
 			},
 			control: { type: 'text' },
 		},
+		ctaIcon: {
+			desciption: 'The `ctaIcon` prop specifies the icon to render in the CTA. Takes an object with `Icon` component props or a string.',
+			table: {
+				type: {
+					summary: 'string',
+				},
+				defaultValue: { summary: 'bag' },
+			},
+			control: {
+				type: 'select',
+				options: [...Object.keys(iconPaths)],
+			},
+		},
 		ctaButtonSuccessText: {
 			description: 'text to temporarily render in the add to cart button after it is clicked',
 			table: {
@@ -230,6 +254,20 @@ export default {
 					summary: 'component',
 				},
 			},
+		},
+		lazyRender: {
+			description: 'Lazy render settings object',
+			defaultValue: {
+				enabled: true,
+				offset: '10%',
+			},
+			table: {
+				type: {
+					summary: 'object',
+				},
+				defaultValue: { summary: 'Lazy render settings object' },
+			},
+			control: { type: 'object' },
 		},
 		breakpoints: {
 			defaultValue: undefined,