+++ packages/snap-store-mobx/src/Search/Stores/SearchPaginationStore.test.ts
@@ -131,6 +131,40 @@ describe('Pagination Store', () => {
 		}
 	});
 
+	it('doesnt allow pageSizeOptions over the api limit of 100', () => {
+		let customPageSizeOptions = [
+			{
+				label: `show 30`,
+				value: 30,
+			},
+			{
+				label: `show 40`,
+				value: 40,
+			},
+			{
+				label: `show 50`,
+				value: 50,
+			},
+			{
+				label: `show 106`,
+				value: 106,
+			},
+		];
+
+		let tempConfig = { ...searchConfig, settings: { pagination: { pageSizeOptions: customPageSizeOptions } } };
+		const pagination = new SearchPaginationStore({
+			config: tempConfig,
+			services,
+			data: {
+				search: searchData,
+				meta: searchData.meta,
+			},
+		});
+
+		expect(pagination.pageSizeOptions.length).toEqual(customPageSizeOptions.length - 1);
+		expect(pagination.pageSizeOptions.filter((opt) => opt.value == 106)).toHaveLength(0);
+	});
+
 	it('knows the begin number', () => {
 		const pagination = new SearchPaginationStore({
 			config: searchConfig,