+++ packages/snap-store-mobx/src/Search/Stores/SearchResultStore.test.ts
@@ -256,7 +256,7 @@ describe('SearchResultStore', () => {
 
 				const variants = (result as Product).variants;
 
-				expect(variants?.data.length).toStrictEqual(parsedVariantData.length);
+				expect(variants?.data.length).toStrictEqual(parsedVariantData.filter((variant: any) => variant.attributes.available !== false).length);
 				expect(variants?.selections.length).toBe(Object.keys(parsedVariantData[0].options).length);
 			});
 		});
@@ -299,14 +299,18 @@ describe('SearchResultStore', () => {
 
 				expect(variants).toBeDefined();
 
-				expect((result as Product).variants?.data.length).toStrictEqual(parsedVariantData.length);
+				expect((result as Product).variants?.data.length).toStrictEqual(
+					parsedVariantData.filter((variant: any) => variant.attributes.available !== false).length
+				);
 				expect((result as Product).variants?.selections.length).toBe(Object.keys(parsedVariantData[0].options).length);
 
 				(result as Product).variants?.update(parsedVariantDataToUse);
 
 				expect((result as Product).variants).toBeDefined();
 
-				expect((result as Product).variants?.data.length).toStrictEqual(parsedVariantDataToUse.length);
+				expect((result as Product).variants?.data.length).toStrictEqual(
+					parsedVariantDataToUse.filter((variant: any) => variant.attributes.available !== false).length
+				);
 				expect((result as Product).variants?.selections.length).toBe(Object.keys(parsedVariantDataToUse[0].options).length);
 			});
 		});
@@ -408,14 +412,18 @@ describe('SearchResultStore', () => {
 
 			expect(variants).toBeDefined();
 
-			expect((resultForTest as Product).variants?.data.length).toStrictEqual(parsedVariantData.length);
+			expect((resultForTest as Product).variants?.data.length).toStrictEqual(
+				parsedVariantData.filter((variant: any) => variant.attributes.available !== false).length
+			);
 			expect((resultForTest as Product).variants?.selections.length).toBe(Object.keys(parsedVariantData[0].options).length);
 
 			(resultForTest as Product).variants?.update(parsedVariantDataToUse);
 
 			expect((resultForTest as Product).variants).toBeDefined();
 
-			expect((resultForTest as Product).variants?.data.length).toStrictEqual(parsedVariantDataToUse.length);
+			expect((resultForTest as Product).variants?.data.length).toStrictEqual(
+				parsedVariantData.filter((variant: any) => variant.attributes.available !== false).length
+			);
 			expect((resultForTest as Product).variants?.selections.length).toBe(Object.keys(parsedVariantDataToUse[0].options).length);
 
 			expect(resultForTest).toBeDefined();
@@ -482,14 +490,18 @@ describe('SearchResultStore', () => {
 
 			expect(variants).toBeDefined();
 
-			expect((resultForTest as Product).variants?.data.length).toStrictEqual(parsedVariantData.length);
+			expect((resultForTest as Product).variants?.data.length).toStrictEqual(
+				parsedVariantData.filter((variant: any) => variant.attributes.available !== false).length
+			);
 			expect((resultForTest as Product).variants?.selections.length).toBe(Object.keys(parsedVariantData[0].options).length);
 
 			(resultForTest as Product).variants?.update(parsedVariantDataToUse);
 
 			expect((resultForTest as Product).variants).toBeDefined();
 
-			expect((resultForTest as Product).variants?.data.length).toStrictEqual(parsedVariantDataToUse.length);
+			expect((resultForTest as Product).variants?.data.length).toStrictEqual(
+				parsedVariantDataToUse.filter((variant: any) => variant.attributes.available !== false).length
+			);
 			expect((resultForTest as Product).variants?.selections.length).toBe(Object.keys(parsedVariantDataToUse[0].options).length);
 
 			expect(resultForTest).toBeDefined();
@@ -525,14 +537,18 @@ describe('SearchResultStore', () => {
 
 			expect(variants).toBeDefined();
 
-			expect((resultForTest as Product).variants?.data.length).toStrictEqual(parsedVariantData.length);
+			expect((resultForTest as Product).variants?.data.length).toStrictEqual(
+				parsedVariantData.filter((variant: any) => variant.attributes.available !== false).length
+			);
 			expect((resultForTest as Product).variants?.selections.length).toBe(Object.keys(parsedVariantData[0].options).length);
 
 			(resultForTest as Product).variants?.update(parsedVariantDataToUse, newVariantsConfig);
 
 			expect((resultForTest as Product).variants).toBeDefined();
 
-			expect((resultForTest as Product).variants?.data.length).toStrictEqual(parsedVariantDataToUse.length);
+			expect((resultForTest as Product).variants?.data.length).toStrictEqual(
+				parsedVariantDataToUse.filter((variant: any) => variant.attributes.available !== false).length
+			);
 			expect((resultForTest as Product).variants?.selections.length).toBe(Object.keys(parsedVariantDataToUse[0].options).length);
 
 			expect(resultForTest).toBeDefined();
@@ -548,6 +564,282 @@ describe('SearchResultStore', () => {
 
 			expect(newcolorSettings).toContain(newcolorSelection?.selected?.value?.toLocaleLowerCase());
 			expect(newinseamSettings).toContain(newinseamSelection?.selected?.value?.toLocaleLowerCase());
+		});
+
+		it('grabs selections from dom on load when using realtime', () => {
+			const field = 'color';
+			const value = 'Mirage';
+
+			//add pdp variant option elems to the dom
+			let colorOptionElem = document.createElement('div');
+			colorOptionElem.setAttribute('ss-variant-option', `${field}:${value}`);
+			colorOptionElem.setAttribute('ss-variant-option-selected', 'true');
+			document.body.appendChild(colorOptionElem);
+
+			const searchData = mockData.updateConfig({ siteId: 'z7h1jh' }).searchMeta('variants');
+
+			const variantSearchConfig: StoreConfigs = {
+				...searchConfig,
+				settings: {
+					variants: {
+						field: 'ss_variants',
+						realtime: {
+							enabled: true,
+						},
+					},
+				},
+			};
+
+			const results = new SearchResultStore({
+				config: variantSearchConfig,
+				state: {
+					loaded: false,
+				},
+				data: {
+					search: searchData,
+					meta: searchData.meta,
+				},
+			});
+
+			expect(results.length).toBe(searchData.pagination?.pageSize);
+
+			const resultForTest = results[0] as Product;
+			expect(resultForTest).toBeDefined();
+
+			const selection = resultForTest.variants?.selections.find((selection) => selection.field == field);
+			expect(selection).toBeDefined();
+			expect(selection?.selected?.value).toBe(value);
+
+			document.body.removeChild(colorOptionElem);
+		});
+
+		it('updates selections from dom elem onclicks realtime', () => {
+			const field = 'color';
+			const value = 'Mirage';
+
+			//add pdp variant option elems to the dom
+			let colorOptionElem = document.createElement('div');
+			colorOptionElem.setAttribute('ss-variant-option', `${field}:${value}`);
+			document.body.appendChild(colorOptionElem);
+
+			const searchData = mockData.updateConfig({ siteId: 'z7h1jh' }).searchMeta('variants');
+
+			const variantSearchConfig: StoreConfigs = {
+				...searchConfig,
+				settings: {
+					variants: {
+						field: 'ss_variants',
+						realtime: {
+							enabled: true,
+						},
+					},
+				},
+			};
+
+			const results = new SearchResultStore({
+				config: variantSearchConfig,
+				state: {
+					loaded: false,
+				},
+				data: {
+					search: searchData,
+					meta: searchData.meta,
+				},
+			});
+
+			expect(results.length).toBe(searchData.pagination?.pageSize);
+
+			const resultForTest = results[0] as Product;
+			expect(resultForTest).toBeDefined();
+
+			let selection = resultForTest.variants?.selections.find((selection) => selection.field == field);
+
+			expect(selection).toBeDefined();
+			expect(selection?.selected?.value).not.toBe(value);
+
+			colorOptionElem.click();
+
+			selection = resultForTest.variants?.selections.find((selection) => selection.field == field);
+
+			expect(selection).toBeDefined();
+			expect(selection?.selected?.value).toBe(value);
+
+			document.body.removeChild(colorOptionElem);
+		});
+
+		it('can use filter first when grabbing selections from dom on load', () => {
+			const field = 'color';
+			const value = 'Mirage';
+
+			//add pdp variant option elems to the dom
+			let colorOptionElem = document.createElement('div');
+			colorOptionElem.setAttribute('ss-variant-option', `${field}:${value}`);
+			colorOptionElem.setAttribute('ss-variant-option-selected', 'true');
+			document.body.appendChild(colorOptionElem);
+
+			const searchData = mockData.updateConfig({ siteId: 'z7h1jh' }).searchMeta('variants');
+
+			const variantSearchConfig: StoreConfigs = {
+				...searchConfig,
+				settings: {
+					variants: {
+						field: 'ss_variants',
+						realtime: {
+							enabled: true,
+							filters: ['first'],
+						},
+					},
+				},
+			};
+
+			const results = new SearchResultStore({
+				config: variantSearchConfig,
+				state: {
+					loaded: false,
+				},
+				data: {
+					search: searchData,
+					meta: searchData.meta,
+				},
+			});
+
+			expect(results.length).toBe(searchData.pagination?.pageSize);
+
+			const resultForTest = results[0] as Product;
+			expect(resultForTest).toBeDefined();
+
+			let selection = resultForTest.variants?.selections.find((selection) => selection.field == field);
+
+			expect(selection?.selected?.value).toBe(value);
+
+			const secondResultForTest = results[1] as Product;
+			expect(secondResultForTest).toBeDefined();
+
+			selection = secondResultForTest.variants?.selections.find((selection) => selection.field == field);
+			expect(selection?.values.some((val) => val.value == value)).toBeTruthy();
+
+			expect(selection?.selected?.value).not.toBe(value);
+
+			document.body.removeChild(colorOptionElem);
+		});
+
+		it('can use filter first while updating selections from dom elem onclicks realtime', () => {
+			const field = 'color';
+			const value = 'Mirage';
+
+			//add pdp variant option elems to the dom
+			let colorOptionElem = document.createElement('div');
+			colorOptionElem.setAttribute('ss-variant-option', `${field}:${value}`);
+			document.body.appendChild(colorOptionElem);
+
+			const searchData = mockData.updateConfig({ siteId: 'z7h1jh' }).searchMeta('variants');
+
+			const variantSearchConfig: StoreConfigs = {
+				...searchConfig,
+				settings: {
+					variants: {
+						field: 'ss_variants',
+						realtime: {
+							enabled: true,
+							filters: ['first'],
+						},
+					},
+				},
+			};
+
+			const results = new SearchResultStore({
+				config: variantSearchConfig,
+				state: {
+					loaded: false,
+				},
+				data: {
+					search: searchData,
+					meta: searchData.meta,
+				},
+			});
+
+			expect(results.length).toBe(searchData.pagination?.pageSize);
+
+			const resultForTest = results[0] as Product;
+			expect(resultForTest).toBeDefined();
+
+			const secondResultForTest = results[1] as Product;
+			expect(secondResultForTest).toBeDefined();
+
+			let selection = resultForTest.variants?.selections.find((selection) => selection.field == field);
+			const selection2 = secondResultForTest.variants?.selections.find((selection) => selection.field == field);
+
+			expect(selection?.selected?.value).not.toBe(value);
+
+			colorOptionElem.click();
+
+			expect(selection?.selected?.value).toBe(value);
+
+			expect(selection2?.values.some((val) => val.value == value)).toBeTruthy();
+			expect(selection2?.selected?.value).not.toBe(value);
+
+			document.body.removeChild(colorOptionElem);
+		});
+
+		it('can use filter unaltered while updating selections from dom elem onclicks realtime', () => {
+			const field = 'color';
+			const value = 'Mirage';
+
+			//add pdp variant option elems to the dom
+			let colorOptionElem = document.createElement('div');
+			colorOptionElem.setAttribute('ss-variant-option', `${field}:${value}`);
+			document.body.appendChild(colorOptionElem);
+
+			const searchData = mockData.updateConfig({ siteId: 'z7h1jh' }).searchMeta('variants');
+
+			const variantSearchConfig: StoreConfigs = {
+				...searchConfig,
+				settings: {
+					variants: {
+						field: 'ss_variants',
+						realtime: {
+							enabled: true,
+							filters: ['unaltered'],
+						},
+					},
+				},
+			};
+
+			const results = new SearchResultStore({
+				config: variantSearchConfig,
+				state: {
+					loaded: false,
+				},
+				data: {
+					search: searchData,
+					meta: searchData.meta,
+				},
+			});
+
+			expect(results.length).toBe(searchData.pagination?.pageSize);
+
+			const resultForTest = results[0] as Product;
+			expect(resultForTest).toBeDefined();
+
+			const secondResultForTest = results[1] as Product;
+			expect(secondResultForTest).toBeDefined();
+
+			let selection = resultForTest.variants?.selections.find((selection) => selection.field == field);
+			const selection2 = secondResultForTest.variants?.selections.find((selection) => selection.field == field);
+
+			expect(selection?.selected?.value).not.toBe(value);
+			expect(selection2?.selected?.value).not.toBe(value);
+
+			selection?.select('Desert');
+
+			colorOptionElem.click();
+
+			expect(selection?.selected?.value).toBe('Desert');
+
+			expect(selection2?.values.some((val) => val.value == value)).toBeTruthy();
+			expect(selection2?.selected?.value).toBe(value);
+
+			document.body.removeChild(colorOptionElem);
 		});
 
 		it('can use the "thumbnailBackgroundImages" option to set the backgroundImageUrl for each variant to the variant thumbnailImageUrl', () => {
@@ -694,14 +986,16 @@ describe('SearchResultStore', () => {
 				expect(variants).toHaveProperty('makeSelections');
 				expect(variants).toHaveProperty('update');
 
+				const filteredParsedVariantsData = parsedVariantData.filter((variant: any) => variant.attributes.available !== false);
+
 				// only uses "available" variants
 				expect(variants?.active).toBe(variants?.data.find((variant) => variant.available));
-				expect(variants?.data.length).toStrictEqual(parsedVariantData.length);
+				expect(variants?.data.length).toStrictEqual(filteredParsedVariantsData.length);
 				expect(variants?.selections.length).toBe(Object.keys(parsedVariantData[0].options).length);
 
-				// creates a variant for each data entry
+				// creates a variant for each available data entry
 				variants.data.forEach((variant, index) => {
-					expect(variant).toStrictEqual(new Variant({ data: { variant: parsedVariantData[index] } }));
+					expect(variant).toStrictEqual(new Variant({ data: { variant: filteredParsedVariantsData[index] } }));
 				});
 			});