+++ packages/snap-store-mobx/src/Search/Stores/SearchResultStore.ts
@@ -147,6 +147,7 @@ export type VariantDataOptions = Record<
 		backgroundImageUrl?: string;
 		attributeId?: string;
 		optionId?: string;
+		optionValue?: string;
 	}
 >;
 
@@ -381,17 +382,29 @@ export class Variants {
 			const options: string[] = [];
 
 			// create variants objects
-			this.data = variantData.map((variant) => {
-				Object.keys(variant.options).forEach((variantOption) => {
-					if (!options.includes(variantOption)) {
-						options.push(variantOption);
+			this.data = variantData
+				.filter((variant) => variant.attributes.available !== false)
+				.map((variant) => {
+					// normalize price fields ensuring they are numbers
+					if (variant.mappings.core?.price) {
+						variant.mappings.core.price = Number(variant.mappings.core?.price);
 					}
-				});
+					if (variant.mappings.core?.msrp) {
+						variant.mappings.core.msrp = Number(variant.mappings.core?.msrp);
+					}
+					return variant;
+				})
+				.map((variant) => {
+					Object.keys(variant.options).forEach((variantOption) => {
+						if (!options.includes(variantOption)) {
+							options.push(variantOption);
+						}
+					});
 
-				return new Variant({
-					data: { variant },
+					return new Variant({
+						data: { variant },
+					});
 				});
-			});
 
 			//need to reset this.selections first
 			this.selections = [];
@@ -555,7 +568,7 @@ export class VariantSelection {
 		// current selection should only consider OTHER selections for availability
 		const selectedSelections = variants.selections.filter((selection) => selection.field != this.field && selection.selected);
 
-		let availableVariants = variants.data;
+		let availableVariants = variants.data.filter((variant) => variant.available);
 
 		// loop through selectedSelections and remove products that do not match
 		for (const selectedSelection of selectedSelections) {