+++ packages/snap-tracker/src/Tracker.test.ts
@@ -88,7 +88,8 @@ describe('Script Block Tracking', () => {
 	it('can target track/order/transaction', async () => {
 		const order = {
 			id: '123456',
-			total: '9.99',
+			total: '10.71',
+			transactionTotal: '9.99',
 			city: 'Los Angeles',
 			state: 'CA',
 			country: 'US',
@@ -479,6 +480,7 @@ describe('Tracker', () => {
 
 		const customGlobals = {
 			siteId: 'custom',
+			currency: { code: 'EUR' },
 		};
 
 		const tracker = new Tracker(globals);
@@ -495,6 +497,9 @@ describe('Tracker', () => {
 		// @ts-ignore - private property
 		expect(tracker.config.framework).toStrictEqual('snap');
 
+		// @ts-ignore - private property
+		expect(tracker.context.currency).toBeUndefined();
+
 		const tracker2 = new Tracker(customGlobals, customConfig);
 
 		// @ts-ignore - private property
@@ -508,6 +513,12 @@ describe('Tracker', () => {
 
 		// @ts-ignore - private property
 		expect(tracker2.config.framework).toBe(customConfig.framework);
+
+		// @ts-ignore - private property
+		expect(tracker2.context.currency).toBeDefined();
+
+		// @ts-ignore - private property
+		expect(tracker2.context.currency).toStrictEqual(customGlobals.currency);
 	});
 
 	it('can persist userId in storage if cookies are disabled', async () => {
@@ -780,14 +791,16 @@ describe('Tracker', () => {
 		const eventFn = jest.spyOn(tracker.track.product, 'view');
 
 		const payload = {
+			uid: undefined,
 			sku: 'abc123',
+			childUid: undefined,
 			childSku: 'abc123_a',
 		};
 		const beaconEvent = await tracker.track.product.view(payload);
 
 		expect(beaconEvent?.type).toStrictEqual(BeaconType.PRODUCT);
 		expect(beaconEvent?.category).toStrictEqual(BeaconCategory.PAGEVIEW);
-		expect(beaconEvent?.event).toStrictEqual(payload);
+		expect(beaconEvent?.event).toEqual(payload);
 
 		expect(eventFn).toHaveBeenCalledTimes(1);
 		expect(eventFn).toHaveBeenCalledWith(payload);
@@ -826,7 +839,7 @@ describe('Tracker', () => {
 		productView.mockRestore();
 	});
 
-	it('logs console error if no sku or childSku is provided to track.product.view', async () => {
+	it('logs console error if no uid or sku or childUid or childSku is provided to track.product.view', async () => {
 		const tracker = new Tracker(globals, config);
 		const eventFn = jest.spyOn(tracker.track.product, 'view');
 		const consoleError = jest.spyOn(console, 'error');
@@ -1026,50 +1039,65 @@ describe('Tracker', () => {
 		eventFn.mockRestore();
 	});
 
-	it('logs console error if no items is provided to track.cart.view', async () => {
+	it('can invoke track.cart.view event method with uid', async () => {
 		const tracker = new Tracker(globals, config);
+
 		const eventFn = jest.spyOn(tracker.track.cart, 'view');
-		const consoleError = jest.spyOn(console, 'error');
 
-		let consoleCount = 0;
-		let payload: CartViewEvent = {
-			items: [],
-		};
-		await tracker.track.cart.view(payload);
-		consoleCount++;
-
-		expect(eventFn).toHaveBeenCalledWith(payload);
-		expect(consoleError).toHaveBeenCalledTimes(consoleCount);
-
-		// without qty, price, childSku or sku
-		payload = {
+		const payload = {
 			items: [
-				// @ts-ignore
 				{
-					childSku: 'abc123_a',
-					// qty: '1',
+					uid: 'id_1',
+					qty: '1',
 					price: '9.99',
 				},
-				// @ts-ignore
 				{
-					childSku: 'def456_a',
-					qty: '2',
-					// price: '10.99',
-				},
-				{
-					// childSku: 'def456_a',
+					uid: 'id_2',
 					qty: '2',
 					price: '10.99',
 				},
 			],
 		};
-		await tracker.track.cart.view(payload);
-		consoleCount += payload.items.length;
+		const beaconEvent = await tracker.track.cart.view(payload);
 
+		expect(beaconEvent?.type).toStrictEqual(BeaconType.CART);
+		expect(beaconEvent?.category).toStrictEqual(BeaconCategory.CARTVIEW);
+		expect(beaconEvent?.event).toStrictEqual(payload);
+
+		expect(eventFn).toHaveBeenCalledTimes(1);
 		expect(eventFn).toHaveBeenCalledWith(payload);
-		expect(consoleError).toHaveBeenCalledTimes(consoleCount);
 
-		consoleError.mockRestore();
+		eventFn.mockRestore();
+	});
+
+	it('can invoke track.cart.view event method with childUid', async () => {
+		const tracker = new Tracker(globals, config);
+
+		const eventFn = jest.spyOn(tracker.track.cart, 'view');
+
+		const payload = {
+			items: [
+				{
+					childUid: 'id_c1',
+					qty: '1',
+					price: '9.99',
+				},
+				{
+					childUid: 'id_c2',
+					qty: '2',
+					price: '10.99',
+				},
+			],
+		};
+		const beaconEvent = await tracker.track.cart.view(payload);
+
+		expect(beaconEvent?.type).toStrictEqual(BeaconType.CART);
+		expect(beaconEvent?.category).toStrictEqual(BeaconCategory.CARTVIEW);
+		expect(beaconEvent?.event).toStrictEqual(payload);
+
+		expect(eventFn).toHaveBeenCalledTimes(1);
+		expect(eventFn).toHaveBeenCalledWith(payload);
+
 		eventFn.mockRestore();
 	});
 
@@ -1114,6 +1142,65 @@ describe('Tracker', () => {
 		trackEvent.mockRestore();
 	});
 
+	it('logs console error if no valid items are provided to track.cart.view', async () => {
+		const tracker = new Tracker(globals, config);
+		const eventFn = jest.spyOn(tracker.track.cart, 'view');
+		const consoleError = jest.spyOn(console, 'error');
+
+		let consoleCount = 0;
+		let payload: CartViewEvent = {
+			items: [],
+		};
+		await tracker.track.cart.view(payload);
+		consoleCount++;
+
+		expect(eventFn).toHaveBeenCalledWith(payload);
+		expect(consoleError).toHaveBeenCalledTimes(consoleCount);
+
+		// without qty, price, childSku or sku
+		payload = {
+			items: [
+				// @ts-ignore
+				{
+					childSku: 'abc123_a',
+					// qty: '1',
+					price: '9.99',
+				},
+				// @ts-ignore
+				{
+					childSku: 'def456_a',
+					qty: '2',
+					// price: '10.99',
+				},
+				// @ts-ignore
+				{
+					childUid: '456_a',
+					// qty: '2',
+					price: '10.99',
+				},
+				// @ts-ignore
+				{
+					uid: '456',
+					// qty: '3',
+					price: '10.99',
+				},
+				{
+					// childSku: 'def456_a',
+					qty: '2',
+					price: '10.99',
+				},
+			],
+		};
+		await tracker.track.cart.view(payload);
+		consoleCount += payload.items.length;
+
+		expect(eventFn).toHaveBeenCalledWith(payload);
+		expect(consoleError).toHaveBeenCalledTimes(consoleCount);
+
+		consoleError.mockRestore();
+		eventFn.mockRestore();
+	});
+
 	it('can invoke track.order.transaction event method', async () => {
 		const tracker = new Tracker(globals, config);
 
@@ -1122,14 +1209,16 @@ describe('Tracker', () => {
 		const payload = {
 			order: {
 				id: '123456',
-				total: '9.99',
+				total: '10.71',
+				transactionTotal: '9.99',
 				city: 'Los Angeles',
 				state: 'CA',
 				country: 'US',
 			},
 			items: [
 				{
-					sku: 'abc123',
+					uid: 'abc123',
+					sku: 'productabc123',
 					childSku: 'abc123_a',
 					qty: '1',
 					price: '9.99',
@@ -1143,6 +1232,7 @@ describe('Tracker', () => {
 		expect(beaconEvent?.event).toStrictEqual({
 			orderId: payload.order.id,
 			total: payload.order.total,
+			transactionTotal: payload.order.transactionTotal,
 			city: payload.order.city,
 			state: payload.order.state,
 			country: payload.order.country,
@@ -1155,7 +1245,7 @@ describe('Tracker', () => {
 		eventFn.mockRestore();
 	});
 
-	it('can invoke track.order.transaction event method without item skus', async () => {
+	it('can invoke track.order.transaction event method with item skus', async () => {
 		const tracker = new Tracker(globals, config);
 
 		const eventFn = jest.spyOn(tracker.track.order, 'transaction');
@@ -1163,7 +1253,8 @@ describe('Tracker', () => {
 		const payload = {
 			order: {
 				id: '123456',
-				total: '9.99',
+				total: '10.71',
+				transactionTotal: '9.99',
 				city: 'Los Angeles',
 				state: 'CA',
 				country: 'US',
@@ -1184,6 +1275,91 @@ describe('Tracker', () => {
 		expect(beaconEvent?.event).toStrictEqual({
 			orderId: payload.order.id,
 			total: payload.order.total,
+			transactionTotal: payload.order.transactionTotal,
+			city: payload.order.city,
+			state: payload.order.state,
+			country: payload.order.country,
+			items: payload.items,
+		});
+
+		expect(eventFn).toHaveBeenCalledTimes(1);
+		expect(eventFn).toHaveBeenCalledWith(payload);
+
+		eventFn.mockRestore();
+	});
+
+	it('can invoke track.order.transaction event method with ONLY item uid', async () => {
+		const tracker = new Tracker(globals, config);
+
+		const eventFn = jest.spyOn(tracker.track.order, 'transaction');
+
+		const payload = {
+			order: {
+				id: '123456',
+				total: '10.71',
+				transactionTotal: '9.99',
+				city: 'Los Angeles',
+				state: 'CA',
+				country: 'US',
+			},
+			items: [
+				{
+					uid: 'abc123_id',
+					qty: '1',
+					price: '9.99',
+				},
+			],
+		};
+		const beaconEvent = await tracker.track.order.transaction(payload);
+
+		expect(beaconEvent?.type).toStrictEqual(BeaconType.ORDER);
+		expect(beaconEvent?.category).toStrictEqual(BeaconCategory.ORDERVIEW);
+		expect(beaconEvent?.event).toStrictEqual({
+			orderId: payload.order.id,
+			total: payload.order.total,
+			transactionTotal: payload.order.transactionTotal,
+			city: payload.order.city,
+			state: payload.order.state,
+			country: payload.order.country,
+			items: payload.items,
+		});
+
+		expect(eventFn).toHaveBeenCalledTimes(1);
+		expect(eventFn).toHaveBeenCalledWith(payload);
+
+		eventFn.mockRestore();
+	});
+
+	it('can invoke track.order.transaction event method with ONLY item childUid', async () => {
+		const tracker = new Tracker(globals, config);
+
+		const eventFn = jest.spyOn(tracker.track.order, 'transaction');
+
+		const payload = {
+			order: {
+				id: '123456',
+				total: '10.71',
+				transactionTotal: '9.99',
+				city: 'Los Angeles',
+				state: 'CA',
+				country: 'US',
+			},
+			items: [
+				{
+					childUid: 'abc123_cid',
+					qty: '1',
+					price: '9.99',
+				},
+			],
+		};
+		const beaconEvent = await tracker.track.order.transaction(payload);
+
+		expect(beaconEvent?.type).toStrictEqual(BeaconType.ORDER);
+		expect(beaconEvent?.category).toStrictEqual(BeaconCategory.ORDERVIEW);
+		expect(beaconEvent?.event).toStrictEqual({
+			orderId: payload.order.id,
+			total: payload.order.total,
+			transactionTotal: payload.order.transactionTotal,
 			city: payload.order.city,
 			state: payload.order.state,
 			country: payload.order.country,
@@ -1205,7 +1381,8 @@ describe('Tracker', () => {
 		let payload: OrderTransactionData = {
 			order: {
 				id: '123456',
-				total: '9.99',
+				total: '10.71',
+				transactionTotal: '9.99',
 				city: 'Los Angeles',
 				state: 'CA',
 				country: 'US',
@@ -1223,7 +1400,8 @@ describe('Tracker', () => {
 		payload = {
 			order: {
 				id: '123456',
-				total: '9.99',
+				total: '10.71',
+				transactionTotal: '9.99',
 				city: 'Los Angeles',
 				state: 'CA',
 				country: 'US',
@@ -1268,10 +1446,12 @@ describe('Tracker', () => {
 		const orderTransaction = jest.spyOn(tracker.track.order, 'transaction');
 
 		const siteId = 'xxxxxx';
+		const currency = 'EUR';
 		const payload = {
 			order: {
 				id: '123456',
-				total: '9.99',
+				total: '10.71',
+				transactionTotal: '9.99',
 				city: 'Los Angeles',
 				state: 'CA',
 				country: 'US',
@@ -1301,6 +1481,7 @@ describe('Tracker', () => {
 			event: {
 				orderId: payload.order.id,
 				total: payload.order.total,
+				transactionTotal: payload.order.transactionTotal,
 				city: payload.order.city,
 				state: payload.order.state,
 				country: payload.order.country,