+++ packages/snap-tracker/src/Tracker.ts
@@ -21,10 +21,11 @@ import {
 	ShopperLoginEvent,
 	TrackErrorEvent,
 	OrderTransactionData,
-	Product,
+	ProductData,
 	TrackerConfig,
 	DoNotTrackEntry,
 	PreflightRequestModel,
+	CurrencyContext,
 } from './types';
 
 export const BATCH_TIMEOUT = 200;
@@ -34,6 +35,8 @@ const SHOPPERID_COOKIE_NAME = 'ssShopperId';
 const COOKIE_EXPIRATION = 31536000000; // 1 year
 const VIEWED_COOKIE_EXPIRATION = 220752000000; // 7 years
 const COOKIE_SAMESITE = 'Lax';
+const COOKIE_DOMAIN =
+	(typeof window !== 'undefined' && window.location.hostname && '.' + window.location.hostname.replace(/^www\./, '')) || undefined;
 const SESSIONID_STORAGE_NAME = 'ssSessionIdNamespace';
 const LOCALSTORAGE_BEACON_POOL_NAME = 'ssBeaconPool';
 const CART_PRODUCTS = 'ssCartProducts';
@@ -89,6 +92,10 @@ export class Tracker {
 			},
 		};
 
+		if (this.globals.currency?.code) {
+			this.context.currency = this.globals.currency;
+		}
+
 		if (!window.searchspring?.tracker) {
 			window.searchspring = window.searchspring || {};
 			window.searchspring.tracker = this;
@@ -99,11 +106,12 @@ export class Tracker {
 		setTimeout(() => {
 			this.targeters.push(
 				new DomTargeter([{ selector: 'script[type^="searchspring/track/"]', emptyTarget: false }], (target: any, elem: Element) => {
-					const { item, items, siteId, shopper, order, type } = getContext(
-						['item', 'items', 'siteId', 'shopper', 'order', 'type'],
+					const { item, items, siteId, shopper, order, type, currency } = getContext(
+						['item', 'items', 'siteId', 'shopper', 'order', 'type', 'currency'],
 						elem as HTMLScriptElement
 					);
 
+					this.setCurrency(currency);
 					switch (type) {
 						case 'searchspring/track/shopper/login':
 							this.track.shopper.login(shopper, siteId);
@@ -313,7 +321,7 @@ export class Tracker {
 				const storedShopperId = this.getShopperId();
 				if (storedShopperId != data.id) {
 					// user's logged in id has changed, update shopperId cookie send login event
-					cookies.set(SHOPPERID_COOKIE_NAME, data.id, COOKIE_SAMESITE, COOKIE_EXPIRATION);
+					cookies.set(SHOPPERID_COOKIE_NAME, data.id, COOKIE_SAMESITE, COOKIE_EXPIRATION, COOKIE_DOMAIN);
 					this.context.shopperId = data.id;
 					this.sendPreflight();
 					const payload = {
@@ -331,9 +339,9 @@ export class Tracker {
 		},
 		product: {
 			view: (data: ProductViewEvent, siteId?: string): BeaconEvent | undefined => {
-				if (!data?.sku && !data?.childSku) {
+				if (!data?.uid && !data?.sku && !data?.childUid && !data?.childSku) {
 					console.error(
-						'track.product.view event: requires a valid sku and/or childSku. \nExample: track.product.view({ sku: "product123", childSku: "product123_a" })'
+						'track.product.view event: requires a valid uid, sku and/or childUid, childSku. \nExample: track.product.view({ uid: "123", sku: "product123", childUid: "123_a", childSku: "product123_a" })'
 					);
 					return;
 				}
@@ -352,7 +360,9 @@ export class Tracker {
 					category: BeaconCategory.PAGEVIEW,
 					context,
 					event: {
+						uid: data?.uid ? `${data.uid}` : undefined,
 						sku: data?.sku ? `${data.sku}` : undefined,
+						childUid: data?.childUid ? `${data.childUid}` : undefined,
 						childSku: data?.childSku ? `${data.childSku}` : undefined,
 					},
 				};
@@ -360,11 +370,17 @@ export class Tracker {
 				const event = this.track.event(payload);
 				if (event) {
 					// save recently viewed products to cookie
-					const sku = data?.sku || data?.childSku;
+					const sku = data?.childSku || data?.childUid || data?.sku || data?.uid;
 					if (sku) {
 						const lastViewedProducts = this.cookies.viewed.get();
 						const uniqueCartItems = Array.from(new Set([...lastViewedProducts, sku])).map((item) => item.trim());
-						cookies.set(VIEWED_PRODUCTS, uniqueCartItems.slice(0, MAX_VIEWED_COUNT).join(','), COOKIE_SAMESITE, VIEWED_COOKIE_EXPIRATION);
+						cookies.set(
+							VIEWED_PRODUCTS,
+							uniqueCartItems.slice(0, MAX_VIEWED_COUNT).join(','),
+							COOKIE_SAMESITE,
+							VIEWED_COOKIE_EXPIRATION,
+							COOKIE_DOMAIN
+						);
 						if (!lastViewedProducts.includes(sku)) {
 							this.sendPreflight();
 						}
@@ -377,6 +393,7 @@ export class Tracker {
 							...payload,
 							event: {
 								sku: data.sku,
+								id: data.uid,
 							},
 						});
 					}
@@ -421,7 +438,7 @@ export class Tracker {
 			view: (data: CartViewEvent, siteId?: string): BeaconEvent | undefined => {
 				if (!Array.isArray(data?.items) || !data?.items.length) {
 					console.error(
-						'track.view.cart event: parameter must be an array of cart items. \nExample: track.view.cart({ items: [{ sku: "product123", childSku: "product123_a", qty: "1", price: "9.99" }] })'
+						'track.view.cart event: parameter must be an array of cart items. \nExample: track.view.cart({ items: [{ id: "123", sku: "product123", childSku: "product123_a", qty: "1", price: "9.99" }] })'
 					);
 					return;
 				}
@@ -436,18 +453,24 @@ export class Tracker {
 					});
 				}
 				const items = data.items.map((item, index) => {
-					if (!item?.qty || !item?.price || (!item?.sku && !item?.childSku)) {
+					if (!item?.qty || !item?.price || (!item?.uid && !item?.sku && !item?.childUid && !item?.childSku)) {
 						console.error(
-							`track.view.cart event: item ${item} at index ${index} requires a valid qty, price, and (sku and/or childSku.) \nExample: track.view.cart({ items: [{ sku: "product123", childSku: "product123_a", qty: "1", price: "9.99" }] })`
+							`track.view.cart event: item at index ${index} requires a valid qty, price, and (uid and/or sku and/or childUid and/or childSku.) \nExample: track.view.cart({ items: [{ uid: "123", sku: "product123", childUid: "123_a", childSku: "product123_a", qty: "1", price: "9.99" }] })`
 						);
 						return;
 					}
-					const product: Product = {
+					const product: ProductData = {
 						qty: `${item.qty}`,
 						price: `${item.price}`,
 					};
+					if (item?.uid) {
+						product.uid = `${item.uid}`;
+					}
 					if (item?.sku) {
 						product.sku = `${item.sku}`;
+					}
+					if (item?.childUid) {
+						product.childUid = `${item.childUid}`;
 					}
 					if (item?.childSku) {
 						product.childSku = `${item.childSku}`;
@@ -465,7 +488,7 @@ export class Tracker {
 				if (event) {
 					// save cart items to cookie
 					if (items.length) {
-						const products = items.map((item) => item?.sku || item?.childSku || '').filter((sku) => sku);
+						const products = items.map((item) => item?.childSku || item?.childUid || item?.sku || item?.uid || '').filter((sku) => sku);
 						this.cookies.cart.add(products);
 					}
 					// legacy tracking
@@ -478,7 +501,7 @@ export class Tracker {
 			transaction: (data: OrderTransactionData, siteId?: string): BeaconEvent | undefined => {
 				if (!data?.items || !Array.isArray(data.items) || !data.items.length) {
 					console.error(
-						'track.order.transaction event: object parameter must contain `items` array of cart items. \nExample: order.transaction({ order: { id: "1001", total: "9.99", city: "Los Angeles", state: "CA", country: "US" }, items: [{ sku: "product123", childSku: "product123_a", qty: "1", price: "9.99" }] })'
+						'track.order.transaction event: object parameter must contain `items` array of cart items. \nExample: order.transaction({ order: { id: "1001", total: "10.71", transactionTotal: "9.99", city: "Los Angeles", state: "CA", country: "US" }, items: [{ uid: "123", sku: "product123", childUid: "123_a", childSku: "product123_a", qty: "1", price: "9.99" }] })'
 					);
 					return;
 				}
@@ -493,18 +516,24 @@ export class Tracker {
 					});
 				}
 				const items = data.items.map((item, index) => {
-					if (!item?.qty || !item?.price || (!item?.sku && !item?.childSku)) {
+					if (!item?.qty || !item?.price || (!item?.uid && !item?.sku && !item?.childUid && !item?.childSku)) {
 						console.error(
-							`track.order.transaction event: object parameter \`items\`: item ${item} at index ${index} requires a valid qty, price, and (sku and/or childSku.) \nExample: order.view({ items: [{ sku: "product123", childSku: "product123_a", qty: "1", price: "9.99" }] })`
+							`track.order.transaction event: object parameter \`items\`: item at index ${index} requires a valid qty, price, and (id or sku and/or childSku.) \nExample: order.view({ items: [{ uid: "123", sku: "product123", childUid: "123_a", childSku: "product123_a", qty: "1", price: "9.99" }] })`
 						);
 						return;
 					}
-					const product: Product = {
+					const product: ProductData = {
 						qty: `${item.qty}`,
 						price: `${item.price}`,
 					};
+					if (item?.uid) {
+						product.uid = `${item.uid}`;
+					}
 					if (item?.sku) {
 						product.sku = `${item.sku}`;
+					}
+					if (item?.childUid) {
+						product.childUid = `${item.childUid}`;
 					}
 					if (item?.childSku) {
 						product.childSku = `${item.childSku}`;
@@ -514,6 +543,7 @@ export class Tracker {
 				const eventPayload = {
 					orderId: data?.order?.id ? `${data.order.id}` : undefined,
 					total: data?.order?.total ? `${data.order.total}` : undefined,
+					transactionTotal: data?.order?.transactionTotal ? `${data.order.transactionTotal}` : undefined,
 					city: data?.order?.city ? `${data.order.city}` : undefined,
 					state: data?.order?.state ? `${data.order.state}` : undefined,
 					country: data?.order?.country ? `${data.order.country}` : undefined,
@@ -546,14 +576,21 @@ export class Tracker {
 		}
 	};
 
+	setCurrency = (currency: CurrencyContext): void => {
+		if (!currency?.code) {
+			return;
+		}
+		this.context.currency = currency;
+	};
+
 	getUserId = (): string | undefined | null => {
 		let userId;
 		try {
 			// use cookies if available, fallback to localstorage
 			if (getFlags().cookies()) {
 				userId = cookies.get(LEGACY_USERID_COOKIE_NAME) || cookies.get(USERID_COOKIE_NAME) || uuidv4();
-				cookies.set(USERID_COOKIE_NAME, userId, COOKIE_SAMESITE, COOKIE_EXPIRATION);
-				cookies.set(LEGACY_USERID_COOKIE_NAME, userId, COOKIE_SAMESITE, COOKIE_EXPIRATION);
+				cookies.set(USERID_COOKIE_NAME, userId, COOKIE_SAMESITE, COOKIE_EXPIRATION, COOKIE_DOMAIN);
+				cookies.set(LEGACY_USERID_COOKIE_NAME, userId, COOKIE_SAMESITE, COOKIE_EXPIRATION, COOKIE_DOMAIN);
 			} else if (getFlags().storage()) {
 				userId = window.localStorage.getItem(USERID_COOKIE_NAME) || uuidv4();
 				window.localStorage.setItem(USERID_COOKIE_NAME, userId);
@@ -573,7 +610,7 @@ export class Tracker {
 			try {
 				sessionId = window.sessionStorage.getItem(SESSIONID_STORAGE_NAME) || uuidv4();
 				window.sessionStorage.setItem(SESSIONID_STORAGE_NAME, sessionId);
-				getFlags().cookies() && cookies.set(SESSIONID_STORAGE_NAME, sessionId, COOKIE_SAMESITE, 0); //session cookie
+				getFlags().cookies() && cookies.set(SESSIONID_STORAGE_NAME, sessionId, COOKIE_SAMESITE, 0, COOKIE_DOMAIN); //session cookie
 			} catch (e) {
 				console.error('Failed to persist session id to session storage:', e);
 			}
@@ -582,7 +619,7 @@ export class Tracker {
 			sessionId = cookies.get(SESSIONID_STORAGE_NAME);
 			if (!sessionId) {
 				sessionId = uuidv4();
-				cookies.set(SESSIONID_STORAGE_NAME, sessionId, COOKIE_SAMESITE, 0);
+				cookies.set(SESSIONID_STORAGE_NAME, sessionId, COOKIE_SAMESITE, 0, COOKIE_DOMAIN);
 			}
 		}
 
@@ -653,7 +690,7 @@ export class Tracker {
 				if (items.length) {
 					const cartItems = items.map((item) => item.trim());
 					const uniqueCartItems = Array.from(new Set(cartItems));
-					cookies.set(CART_PRODUCTS, uniqueCartItems.join(','), COOKIE_SAMESITE, 0);
+					cookies.set(CART_PRODUCTS, uniqueCartItems.join(','), COOKIE_SAMESITE, 0, COOKIE_DOMAIN);
 
 					const itemsHaveChanged = cartItems.filter((item) => items.includes(item)).length !== items.length;
 					if (itemsHaveChanged) {
@@ -666,7 +703,7 @@ export class Tracker {
 					const currentCartItems = this.cookies.cart.get();
 					const itemsToAdd = items.map((item) => item.trim());
 					const uniqueCartItems = Array.from(new Set([...currentCartItems, ...itemsToAdd]));
-					cookies.set(CART_PRODUCTS, uniqueCartItems.join(','), COOKIE_SAMESITE, 0);
+					cookies.set(CART_PRODUCTS, uniqueCartItems.join(','), COOKIE_SAMESITE, 0, COOKIE_DOMAIN);
 
 					const itemsHaveChanged = currentCartItems.filter((item) => itemsToAdd.includes(item)).length !== itemsToAdd.length;
 					if (itemsHaveChanged) {
@@ -679,7 +716,7 @@ export class Tracker {
 					const currentCartItems = this.cookies.cart.get();
 					const itemsToRemove = items.map((item) => item.trim());
 					const updatedItems = currentCartItems.filter((item) => !itemsToRemove.includes(item));
-					cookies.set(CART_PRODUCTS, updatedItems.join(','), COOKIE_SAMESITE, 0);
+					cookies.set(CART_PRODUCTS, updatedItems.join(','), COOKIE_SAMESITE, 0, COOKIE_DOMAIN);
 
 					const itemsHaveChanged = currentCartItems.length !== updatedItems.length;
 					if (itemsHaveChanged) {
@@ -689,7 +726,7 @@ export class Tracker {
 			},
 			clear: () => {
 				if (this.cookies.cart.get().length) {
-					cookies.unset(CART_PRODUCTS);
+					cookies.unset(CART_PRODUCTS, COOKIE_DOMAIN);
 					this.sendPreflight();
 				}
 			},