+++ packages/snap-event-manager/CHANGELOG.md
@@ -3,6 +3,38 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [0.59.0](https://github.com/searchspring/snap/compare/v0.58.3...v0.59.0) (2024-08-22)
+
+**Note:** Version bump only for package @searchspring/snap-event-manager
+
+## [0.58.3](https://github.com/searchspring/snap/compare/v0.58.2...v0.58.3) (2024-07-31)
+
+**Note:** Version bump only for package @searchspring/snap-event-manager
+
+## [0.58.2](https://github.com/searchspring/snap/compare/v0.58.1...v0.58.2) (2024-07-08)
+
+**Note:** Version bump only for package @searchspring/snap-event-manager
+
+## [0.58.1](https://github.com/searchspring/snap/compare/v0.58.0...v0.58.1) (2024-07-03)
+
+**Note:** Version bump only for package @searchspring/snap-event-manager
+
+# [0.58.0](https://github.com/searchspring/snap/compare/v0.57.0...v0.58.0) (2024-07-02)
+
+**Note:** Version bump only for package @searchspring/snap-event-manager
+
+# [0.57.0](https://github.com/searchspring/snap/compare/v0.56.6...v0.57.0) (2024-06-26)
+
+**Note:** Version bump only for package @searchspring/snap-event-manager
+
+## [0.56.6](https://github.com/searchspring/snap/compare/v0.56.5...v0.56.6) (2024-06-07)
+
+**Note:** Version bump only for package @searchspring/snap-event-manager
+
+## [0.56.5](https://github.com/searchspring/snap/compare/v0.56.4...v0.56.5) (2024-06-03)
+
+**Note:** Version bump only for package @searchspring/snap-event-manager
+
 ## [0.56.4](https://github.com/searchspring/snap/compare/v0.56.3...v0.56.4) (2024-05-28)
 
 **Note:** Version bump only for package @searchspring/snap-event-manager